drop broken cached property pattern, use our own implementation
This was a fun one. So the comment I wrote about that pattern was kinda wrong;
it turns out that while setting the field attribute *after* the @property def
*did* in fact fix the default value issue, it also straight up overwrote the
custom property. It also crashed MyPy. Go figure.

Anyhow, this is a different approach, where we do the cache invalidation in
the cached_property_dep decorator. This is a modified version of cached_property
that stores the cached value inside of itself instead of the instance dict
(which is faster for lookups, but also means that __get__ only gets called
at caching time and never again, meaning we couldn't put our fancy logic there).
I had no idea descriptors existed before I checked this out, so I guess I
learned something today :)